+++ include/votca/xtp/energy_costfunction.h
@@ -52,7 +52,9 @@ class Energy_costfunction : public Optimiser_costfunction {
 
   Eigen::VectorXd EvaluateGradient(const Eigen::VectorXd& parameters) override;
 
-  int NumParameters() const override { return _orbitals.QMAtoms().size() * 3; };
+  int NumParameters() const override {
+    return int(_orbitals.QMAtoms().size() * 3);
+  };
 
   bool Converged(const Eigen::VectorXd& delta_parameters, double delta_cost,
                  const Eigen::VectorXd& gradient) override;
